refactor(frontend): integrate popover into sidebar and reorganize UI components
