+++ apps/frontend/src/app/components/molecules/index.tsx
@@ -1,10 +1,10 @@
-export * from './app-sidebar-menu'
-export * from './app-sidebar-footer'
-export * from './orderSummaryCard'
-export * from './pendingOrderCard'
-export * from './orderListItem'
-export * from './stockManageListItem'
-export * from './orderListManageItem'
+export * from './sidebar/app-sidebar-menu'
+export * from './sidebar/app-sidebar-footer'
+export * from './order/orderSummaryCard'
+export * from './order/pendingOrderCard'
+export * from './order/orderListItem'
+export * from './stock/stockManageListItem'
+export * from './order/orderListManageItem'
 export * from './chart/barChartRenderer'
 export * from './chart/datePicker'
 export * from './chart/dateRangePicker'
