+++ apps/frontend/src/app/components/molecules/sidebar/app-sidebar-footer.tsx
