+++ apps/frontend/src/app/components/molecules/stock/stockManageListItem.tsx
